feat : storageImg 삭제 관련 utils 생성
